Fixed failing tests

+++ contracts/SupplyChain.sol
@@ -5,9 +5,9 @@ contract SupplyChain {
     // <owner>
     address public owner;
     // <skuCount>
-    uint256 public skuCount;
+    uint public skuCount;
     // <items mapping>
-    mapping(uint256 => Item) public items;
+    mapping(uint => Item) public items;
     // <enum State: ForSale, Sold, Shipped, Received>
     enum State {
         ForSale,
@@ -30,14 +30,14 @@ contract SupplyChain {
      */
 
     // <LogForSale event: sku arg>
-    event LogForSale(uint256 sku);
+    event LogForSale(uint sku);
 
     // <LogSold event: sku arg>
-    event LogSold(uint256 sku);
+    event LogSold(uint sku);
     // <LogShipped event: sku arg>
-    event LogShipped(uint256 sku);
+    event LogShipped(uint sku);
     // <LogReceived event: sku arg>
-    event LogReceived(uint256 sku);
+    event LogReceived(uint sku);
 
     /*
      * Modifiers
@@ -56,16 +56,16 @@ contract SupplyChain {
         _;
     }
 
-    modifier paidEnough(uint256 _price) {
+    modifier paidEnough(uint _price) {
         require(msg.value >= _price);
         _;
     }
 
-    modifier checkValue(uint256 _sku) {
+    modifier checkValue(uint _sku) {
         //refund them after pay for item (why it is before, _ checks for logic before func)
         _;
-        uint256 _price = items[_sku].price;
-        uint256 amountToRefund = msg.value - _price;
+        uint _price = items[_sku].price;
+        uint amountToRefund = msg.value - _price;
         items[_sku].buyer.transfer(amountToRefund);
     }
 
@@ -77,22 +77,22 @@ contract SupplyChain {
     // that an Item is for sale. Hint: What item properties will be non-zero when
     // an Item has been added?
 
-    modifier forSale(uint256 _sku) {
+    modifier forSale(uint _sku) {
         require(
             items[_sku].state == State.ForSale &&
                 items[_sku].seller != address(0)
         );
         _;
     }
-    modifier sold(uint256 _sku) {
+    modifier sold(uint _sku) {
         require(items[_sku].state == State.Sold);
         _;
     }
-    modifier shipped(uint256 _sku) {
+    modifier shipped(uint _sku) {
         require(items[_sku].state == State.Shipped);
         _;
     }
-    modifier received(uint256 _sku) {
+    modifier received(uint _sku) {
         require(items[_sku].state == State.Received);
         _;
     }
@@ -105,7 +105,7 @@ contract SupplyChain {
         skuCount = 0;
     }
 
-    function addItem(string memory _name, uint256 _price)
+    function addItem(string memory _name, uint _price)
         public
         returns (bool)
     {
@@ -159,7 +159,7 @@ contract SupplyChain {
     //    - the person calling this function is the seller.
     // 2. Change the state of the item to shipped.
     // 3. call the event associated with this function!
-    function shipItem(uint256 sku)
+    function shipItem(uint sku)
         public
         sold(sku)
         verifyCaller(items[sku].seller)
